[変更] HPとAPをintからfloatに変更
* HPゲージはfloatで管理するため
* AP半減を実装する際，少数が発生するため

+++ Assets/Scripts/Breakable.cs
@@ -11,10 +11,10 @@ public class Breakable : MonoBehaviour {
 			{"Invincible", new List<string> {}}
 		};
 
-	public int maxHitPoint;
+	public float maxHitPoint;
 
 	[System.NonSerialized]
-	public int hitPoint;
+	public float hitPoint;
 
 	protected virtual void Start () {
 		hitPoint = maxHitPoint;
